Move make_fixture to utils.py.
